Moves custom activity data to activities table
Primarily because there's not really any evidence that it's warranted
as a separated table, but also because it's not ideal to manage.

There's a chance the extra inserts can be lost, we have to manually
batch them together which is hacky.

Most critically, not every db (sqlite) supports getGeneratedKeys
on batches so it wasn't working regardless.

+++ prism-bukkit/src/main/java/network/darkhelmet/prism/PrismBukkit.java
@@ -156,7 +156,7 @@ public class PrismBukkit implements IPrism {
      * Sets a numeric version we can use to handle differences between serialization formats.
      */
     @Getter
-    protected int serializerVersion;
+    protected short serializerVersion;
 
     /**
      * The storage adapter.
@@ -210,7 +210,7 @@ public class PrismBukkit implements IPrism {
         );
         dependencyService.loadAllDependencies(platformDependencies());
 
-        serializerVersion = DataFixerUtil.getCurrentVersion();
+        serializerVersion = (short) DataFixerUtil.getCurrentVersion();
         bootstrap.loggingService().logger().info(String.format("Serializer version: %d", serializerVersion));
 
         injectorProvider = new InjectorProvider(this, bootstrap.loggingService());
