+++ prism-core/src/main/java/network/darkhelmet/prism/core/storage/adapters/mariadb/MariaDbStorageAdapter.java
@@ -69,7 +69,7 @@ public class MariaDbStorageAdapter extends AbstractSqlStorageAdapter {
             SqlSchemaUpdater schemaUpdater,
             ISqlActivityQueryBuilderFactory queryBuilderFactory,
             CacheService cacheService,
-            @Named("serializerVersion") int serializerVersion,
+            @Named("serializerVersion") short serializerVersion,
             Path dataPath) {
         super(
             loggingService,
