+++ prism-core/src/main/java/network/darkhelmet/prism/core/storage/adapters/mysql/MysqlStorageAdapter.java
@@ -69,7 +69,7 @@ public class MysqlStorageAdapter extends AbstractSqlStorageAdapter {
             SqlSchemaUpdater schemaUpdater,
             ISqlActivityQueryBuilderFactory queryBuilderFactory,
             CacheService cacheService,
-            @Named("serializerVersion") int serializerVersion,
+            @Named("serializerVersion") short serializerVersion,
             Path dataPath) {
         super(
             loggingService,
