+++ prism-core/src/main/java/network/darkhelmet/prism/core/storage/adapters/postgres/PostgresStorageAdapter.java
@@ -72,7 +72,7 @@ public class PostgresStorageAdapter extends AbstractSqlStorageAdapter {
             SqlSchemaUpdater schemaUpdater,
             ISqlActivityQueryBuilderFactory queryBuilderFactory,
             CacheService cacheService,
-            @Named("serializerVersion") int serializerVersion,
+            @Named("serializerVersion") short serializerVersion,
             Path dataPath) {
         super(
             loggingService,
